+++ .gitignore
@@ -2,4 +2,4 @@
 __pycache__/
 *.dot
 *.png
-!/gates/*.png
+!/resources/*.png
