+++ bdd/__init__.py
@@ -1,0 +1,3 @@
+from bdd.bddconstructor import BDDConstructor
+from bdd.bdd import BDD
+from bdd.bddminimiser import BDDMinimiser
