+++ exceptions/__init__.py
@@ -1,0 +1,1 @@
+from exceptions.galileoparseexception import GalileoParseException
