+++ exceptions/galileoparseexception.py
@@ -1,0 +1,7 @@
+class GalileoParseException(Exception):
+    """
+    GalileoParseException is an exception used when the parsing of a
+    Galileo-formatted file failed.
+    """
+
+    pass
