+++ faulttree/gates/__init__.py
@@ -1,0 +1,6 @@
+from faulttree.gates.andgate import AndGate
+from faulttree.gates.basicevent import BasicEvent
+from faulttree.gates.notgate import NotGate
+from faulttree.gates.orgate import OrGate
+from faulttree.gates.votgate import VotGate
+from faulttree.gates.xorgate import XorGate
