+++ faulttree/readers/__init__.py
@@ -1,0 +1,1 @@
+from faulttree.readers.galileoreader import GalileoReader
