+++ resources/andgate.png
