+++ resources/xorgate.png
